pyproject.toml
[tool.poetry]
name = "DermaML"
version = "0.0.0"
description = ""
license = "Apache-2.0"
authors = [
    "April Zhang <april.zhang@berkeley.edu>",
    "Hannah Zhang <hannahzhang@ucsb.edu>",
    "Nicole Tin <nicolelrtin@berkeley.edu>",
    "Kevin Chu <kevin@velexi.com>"
]

[tool.poetry.dependencies]
python = "^3.10,<3.12"

# data management
dvc = {extras = ["s3"], version = "^3"}
fds = {git = "https://github.com/velexi-research/fds.git"}  # Fork of FastDS

# research infrastructure
ipython = "^8"
jupyter-dash = "^0.4"
jupyterlab = "^4"
matplotlib = "^3"
mlflow = "^2"

# miscellaneous utilities
pdoc = "^14"
python-slugify = "^8"
PyYAML = "^6"
tqdm = "^4"
typer = "^0,>=0.12"

# software testing
coverage = "^7"
pytest = "^8"
pytest-cov = "^5"
pytest-flake8 = "^1.1"
pytest-xdist = "^3"

# code quality
black = {extras = ["jupyter"], version = "^24"}
flake8 = "<5.0.0"  # Unpin after compatibility with pytest-flake8 is fixed
pre-commit = "^3"

# security vulnerability fixes
pywin32 = [
    { version = '>=301', platform = 'win32' },
]

# project-specific packages
kaleido = "^0,<=0.2.1"
mediapipe = "^0.10"
opencv-python = "^4"
pillow = "^10"
pycaret = "^3.3.2"
scikit-image = "^0.23"
rembg = "^2"
pyfeats = "^1.0.1"
shap = "^0.46.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --doctest-modules --flake8 --cov --cov-report="
testpaths = [
    "src",
]
doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
    "ALLOW_UNICODE",
    "ALLOW_BYTES",
]
norecursedirs = ["data", "docs", "extras", "notebooks", "reports", "spikes"]

[tool.coverage.run]
source = ["src/"]

[tool.coverage.report]
omit = ["*/tests/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
